+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Validation of component slug ([#153])
 * `component compile` now applies postprocessing filters ([#154])
 * Option to disable postprocessing filters ([#155])
+* `--interactive` option to prompt push confirmation ([#157])
 
 ## [v0.2.2]