+++ CHANGELOG.md
@@ -148,3 +148,4 @@ Initial implementation
 [#153]: https://github.com/projectsyn/commodore/pull/153
 [#154]: https://github.com/projectsyn/commodore/pull/154
 [#155]: https://github.com/projectsyn/commodore/pull/155
+[#157]: https://github.com/projectsyn/commodore/pull/157