+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * `component compile` now applies postprocessing filters ([#154])
 * Option to disable postprocessing filters ([#155])
 * `--interactive` option to prompt push confirmation ([#157])
+* Build Helm bindings for native Helm dependencies ([#161])
 
 ## [v0.2.2]
 
@@ -149,3 +150,4 @@ Initial implementation
 [#154]: https://github.com/projectsyn/commodore/pull/154
 [#155]: https://github.com/projectsyn/commodore/pull/155
 [#157]: https://github.com/projectsyn/commodore/pull/157
+[#161]: https://github.com/projectsyn/commodore/pull/161