+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Extend component template with docs setup ([#158])
 * Build Helm bindings for native Helm dependencies ([#161])
 
+### Changed
+
+* Reworked documentation ([#147])
+
 ## [v0.2.2]
 
 ### Fixed
@@ -146,6 +150,7 @@ Initial implementation
 [#140]: https://github.com/projectsyn/commodore/pull/140
 [#143]: https://github.com/projectsyn/commodore/pull/143
 [#145]: https://github.com/projectsyn/commodore/pull/145
+[#147]: https://github.com/projectsyn/commodore/pull/153
 [#148]: https://github.com/projectsyn/commodore/pull/148
 [#151]: https://github.com/projectsyn/commodore/pull/151
 [#153]: https://github.com/projectsyn/commodore/pull/153