+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Please document all notable changes to this project in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [v0.2.3]
 
 ### Added
 
@@ -129,6 +129,7 @@ Initial implementation
 [v0.2.0]: https://github.com/projectsyn/commodore/releases/tag/v0.2.0
 [v0.2.1]: https://github.com/projectsyn/commodore/releases/tag/v0.2.1
 [v0.2.2]: https://github.com/projectsyn/commodore/releases/tag/v0.2.2
+[v0.2.3]: https://github.com/projectsyn/commodore/releases/tag/v0.2.3
 
 [#53]: https://github.com/projectsyn/commodore/pull/53
 [#58]: https://github.com/projectsyn/commodore/pull/58